requirements.txt
playwrightrequests
